Move tests to /test dir outside of src.
